% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_cond.R
\name{calculate_all}
\alias{calculate_all}
\title{A function to calculate the likelihood of the fed parameters}
\usage{
calculate_all(current_tree_mh, new_k1, pars)
}
\arguments{
\item{current_tree_mh}{The current trees set}

\item{new_k1}{The current value of k2}

\item{pars}{The hyperparameters list}
}
\value{
The resulting value
}
\description{
A function to calculate the likelihood of the fed parameters
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
