% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{calculate_parts_mu}
\alias{calculate_parts_mu}
\title{A function to calculate the terms for the mu posterior}
\usage{
calculate_parts_mu(split, k1l = k1, k2l = k2, Pp = P)
}
\arguments{
\item{split}{The splitted data}

\item{k1l}{The current value of the k1 parameter}

\item{k2l}{The current value of the k2 parameter}

\item{Pp}{The number of trees}
}
\description{
A function to calculate the terms for the mu posterior
}
