% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{calculate_preds}
\alias{calculate_preds}
\title{Calculates predictions}
\usage{
calculate_preds(results, tree, newdata, k1, tau)
}
\arguments{
\item{results}{The results (description to be completed)}

\item{tree}{The latest tree results}

\item{newdata}{The newdata to predict for}

\item{k1}{The latest sampled k1 value}

\item{tau}{The latest sampled tau value}
}
\value{
The sampled values
}
\description{
Calculates predictions for the HEBART model
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
