% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar_functions.R
\name{data_handler}
\alias{data_handler}
\title{A function to adjust the data}
\usage{
data_handler(formula, data, group_variable, scale_fc = FALSE)
}
\arguments{
\item{formula}{The model formula.}

\item{data}{The modelling dataset.}

\item{group_variable}{The grouping variable.}

\item{scale_fc}{Logical to decide whether to scale y or not}
}
\description{
A function to adjust the data
}
\keyword{internal}
