% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnostics}
\alias{diagnostics}
\title{Plots all diagnostics}
\usage{
diagnostics(model)
}
\arguments{
\item{model}{The HEBART model object}
}
\value{
The correspondent plot
}
\description{
Plots all diagnostics plots
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
