% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnostics_density_plot}
\alias{diagnostics_density_plot}
\title{Density plots for tau, k1 or sqrt(k1/tau)}
\usage{
diagnostics_density_plot(model, type = "tau", sqrt = FALSE)
}
\arguments{
\item{model}{The HEBART model object}

\item{type}{Type of plot: tau, k1 or sqrt(k1/tau)}

\item{sqrt}{Logical to decide whether to plot 1/sqrt(tau) instead
(residual precision)}
}
\value{
The correspondent plot
}
\description{
Plots the sampled of values as a density
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
