% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnostics_traceplot}
\alias{diagnostics_traceplot}
\title{Traceplot for tau, k1 or sqrt(k1/tau)}
\usage{
diagnostics_traceplot(model, type = "tau", sqrt = FALSE)
}
\arguments{
\item{model}{The HEBART model object}

\item{type}{Type of plot: tau, k1 or sqrt(k1/tau)}

\item{sqrt}{Logical to decide whether to plot 1/sqrt(tau) instead
(residual precision)}
}
\value{
The correspondent plot
}
\description{
Plots the sampled of values of tau, k1 or sqrt(k1/tau) per iteration
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
