% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_cond.R
\name{faster_det}
\alias{faster_det}
\title{A function to calculate the determinant of M (faster than solve)}
\usage{
faster_det(k_1_d, k_2_d, M_d)
}
\arguments{
\item{k_1_d}{The current value of k1}

\item{k_2_d}{The current value of k2}

\item{M_d}{The M matrix}
}
\value{
The determinant
}
\description{
A function to calculate the determinant of M (faster than solve)
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
