% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{get_avg_nodes}
\alias{get_avg_nodes}
\title{Gets node averages}
\usage{
get_avg_nodes(model)
}
\arguments{
\item{model}{The HEBART model object}
}
\value{
The correspondent averages
}
\description{
Get node averages per tree
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
