% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow_ratio.R
\name{grow_tree}
\alias{grow_tree}
\title{Grows the current tree}
\usage{
grow_tree(current_tree, selec_var, drawn_node, rule)
}
\arguments{
\item{current_tree}{The current tree}

\item{selec_var}{The node to split on node}

\item{drawn_node}{The node to split on}

\item{rule}{The variable value to split on}
}
\value{
The new tree
}
\description{
Grows the current tree
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
