% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hebart.R
\name{hebart}
\alias{hebart}
\title{Hierarchical Embedded Bayesian Additive Regression Trees}
\usage{
hebart(
  formula,
  dataset,
  iter = 100,
  group_variable = "group",
  pars,
  min_u = 0,
  max_u = 20,
  prior_k1 = TRUE,
  num.trees = 5,
  sample_k1 = TRUE,
  burn_in = 50,
  alpha_grow = 0.9,
  beta_grow = 0.5,
  ...
)
}
\arguments{
\item{formula}{The model formula}

\item{dataset}{The data to be used in the modeling}

\item{iter}{The number of MCMC iterations}

\item{group_variable}{The grouping variable}

\item{pars}{The hyperparameters set/list}

\item{min_u}{Integer representing the lower interval of the
Uniform distribution used to sample k1}

\item{max_u}{Integer representing the upper interval of the}

\item{prior_k1}{Logical to decide whether or not use a prior for k1
Uniform distribution used to sample k1}

\item{num.trees}{The number of trees}

\item{sample_k1}{Logical to decide whether to sample_k1 or not}

\item{burn_in}{The number of burn-in iterations}

\item{alpha_grow}{Number between 0 and 1 used in the growing probability
calculation}

\item{beta_grow}{Number between 0 and 1 used in the growing probability
calculation}

\item{...}{Other parameters}
}
\value{
A list containing:
 the sampled values of tau and k1, the final trees
}
\description{
This function runs a BCART model and returns the tree and
other results obtained in the last iteration of the MCMC
}
\details{
Priors used ----------------------------------------------------------
y_{ij} ~ Normal(m_j, tau^-{1})
tau    ~ Gamma(alpha, beta)
mu     ~ Normal(0, tau_mu = k2*tau^-{1})
mu_j   ~ Normal(mu, k1*tau^-{1})
----------------------------------------------------------------------
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
