% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow_ratio.R
\name{inv2}
\alias{inv2}
\alias{cond_calculation}
\title{A function to calculate the inversion of the M matrix}
\usage{
inv2(k_1, k_2, M)

cond_calculation(data_cond, pars)
}
\arguments{
\item{k_1}{The current value of k1}

\item{k_2}{The current value of k2}

\item{M}{The M matrix}

\item{data_cond}{The data to calculate the conditional distribution on}

\item{pars}{The hyperparameters list}
}
\value{
The inverted matrix

The conditional probability
}
\description{
A function to calculate the inversion of the M matrix

Full parent conditional
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.

Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
