% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow_ratio.R
\name{lk_ratio_grow}
\alias{lk_ratio_grow}
\title{Likelihood ratio for grow}
\usage{
lk_ratio_grow(tree, current_node, pars)
}
\arguments{
\item{tree}{The current tree}

\item{current_node}{The current pruned node}

\item{pars}{The full list of parameters}
}
\value{
The likelihood ratio
}
\description{
Likelihood ratio of the candidate tree and the previous
tree, given that the action step was a grow
}
\details{
The likelihood ratio for growth needs the
joint distribution of each region (of each node), given the
tau parameter
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
