% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_cond.R
\name{lk_ratio_k}
\alias{lk_ratio_k}
\title{Likelihood ratio for the sampled k1}
\usage{
lk_ratio_k(new_k1, k1, current_tree, pars, group, tau_post = tau_post, M)
}
\arguments{
\item{new_k1}{The sampled k1}

\item{k1}{The current value of k1}

\item{current_tree}{The current tree}

\item{pars}{The full list of hyperparameters}

\item{group}{The grouping variable}

\item{tau_post}{The current value of tau}

\item{M}{The M matrix}
}
\value{
The likelihood ratio
}
\description{
Likelihood ratio of the candidate tree and the previous
tree, given that the action step was a prune.
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
