% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_ratio.R
\name{lk_ratio_prune}
\alias{lk_ratio_prune}
\title{Likelihood ratio for prune.}
\usage{
lk_ratio_prune(old_tree, tree, current_node, pars, nodes_to_prune)
}
\arguments{
\item{old_tree}{The previous tree.}

\item{tree}{The current tree.}

\item{current_node}{The current pruned node.}

\item{pars}{The full list of parameters.}

\item{nodes_to_prune}{The nodes to prune.}
}
\value{
The likelihood ratio.
}
\description{
Likelihood ratio of the candidate tree and the previous
tree, given that the action step was a prune.
}
\details{
For the likelihood ratio of the new pruned tree, we need
to calculate an inversion of the one in the grow version.
The value is based on the likelihood of all regions, given the
tree and the parameters, for the new and the previous trees.
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
