% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_cond.R
\name{marginal_dist}
\alias{marginal_dist}
\title{A function to calculate the marginal distribution of y}
\usage{
marginal_dist(y, k_1, k_2, M, mu_mu, alpha, beta)
}
\arguments{
\item{y}{The y vector}

\item{k_1}{The current value of k1}

\item{k_2}{The current value of k2}

\item{M}{The M matrix}

\item{mu_mu}{The value for mu_mu (usually 0)}

\item{alpha}{The alpha value from the prior of tau}

\item{beta}{The beta value from the prior of tau}
}
\value{
The resulting value
}
\description{
A function to calculate the marginal distribution of y
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
