% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{MH_update_hbart}
\alias{MH_update_hbart}
\alias{MH_update}
\title{The metropolis-hastings update for k1}
\usage{
MH_update(
  current_tree_mh,
  i = i,
  prior = prior_k1,
  min_u = min_u,
  max_u = max_u,
  pars = pars
)
}
\arguments{
\item{current_tree_mh}{The number of trees}

\item{i}{The current iteration number}

\item{prior}{Logical, to decide whether to use a prior or not}

\item{min_u}{The minimum value of the sampled values for k1}

\item{max_u}{The maximum value of the sampled values for k1}

\item{pars}{The hyperparameters list}
}
\description{
The metropolis-hastings update for k1
}
\details{
If prior = TRUE, this function uses a Weibull(7, 10) prior for k1
}
