% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliar_functions.R
\name{p_rule}
\alias{p_rule}
\title{Rule selection.}
\usage{
p_rule(variable_index, data, sel_node)
}
\arguments{
\item{variable_index}{The variable to create the split.}

\item{data}{The current tree.}

\item{sel_node}{The node to break from.}
}
\value{
The selected splitting value.
}
\description{
Selects a value to split the tree in a grow step.
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
