% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot_avg_nodes}
\alias{plot_avg_nodes}
\title{Plots node averages}
\usage{
plot_avg_nodes(model)
}
\arguments{
\item{model}{The HEBART model object}
}
\value{
The correspondent averages
}
\description{
Gets node averages per tree and plots it
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
