% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot_mse_iter}
\alias{plot_mse_iter}
\title{MSE per iteration}
\usage{
plot_mse_iter(model)
}
\arguments{
\item{model}{The HEBART model object}
}
\value{
The correspondent plot
}
\description{
Plots training MSE values per iteration
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
