% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_hebart}
\alias{predict_hebart}
\title{Predictions for the HEBART model}
\usage{
predict_hebart(model, newdata, formula, group_variable)
}
\arguments{
\item{model}{The model object}

\item{newdata}{The new data to predict}

\item{formula}{The model formula}

\item{group_variable}{The grouping variable}
}
\value{
A dataframe with the prediction column
}
\description{
This function predicts for the final trees of a
Hierachical Embedded BART model
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
