% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.hebart}
\alias{print.hebart}
\title{Print hebart}
\usage{
\method{print}{hebart}(x, ...)
}
\arguments{
\item{x}{Object of class 'hebart'}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Print hebart
}
\seealso{
\code{\link{hebart}}
}
\author{
Bruna Wundervald
}
