% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_ratio.R
\name{prune_tree}
\alias{prune_tree}
\title{Prunes the current tree}
\usage{
prune_tree(current_tree, drawn_node, variable_in_question)
}
\arguments{
\item{current_tree}{The current tree}

\item{drawn_node}{The node to prune}

\item{variable_in_question}{Variable that was split in the node that will}
}
\value{
The new tree
}
\description{
Prunes the current tree
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
