% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriors.R
\name{sample_parameters}
\alias{sample_parameters}
\title{sample_parameters}
\usage{
sample_parameters(
  type = "tau",
  P = P,
  i = i,
  N = NULL,
  beta = NULL,
  alpha = NULL,
  current_tree_post = NULL,
  k1 = NULL,
  k2 = NULL,
  tau_post,
  J = NULL,
  mu_post = NULL
)
}
\arguments{
\item{type}{The type of parameter to be sampled ("tau", "mu" or "muj")}

\item{P}{The number of trees}

\item{i}{The current iteration number}

\item{N}{The size of the data}

\item{beta}{A parameter from the tau prior}

\item{alpha}{A parameter from the tau prior}

\item{current_tree_post}{The current trees object}

\item{k1}{The current value of the k1 parameter}

\item{k2}{The current value of the k2 parameter}

\item{tau_post}{The current value of the posterior tau}

\item{J}{The number of groups}

\item{mu_post}{The current set of mu posterior values}
}
\description{
sample_parameters
}
