% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{sample_pred}
\alias{sample_pred}
\title{Prediction sampling}
\usage{
sample_pred(mu, k1, tau)
}
\arguments{
\item{mu}{The mu value to use as the mean}

\item{k1}{The latest sampled k1 value}

\item{tau}{The latest sampled tau value}
}
\value{
The sampled values
}
\description{
This function predicts for the final trees of a
Hierachical Embedded BART model
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
