% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_ratio.R
\name{structure_ratio_prune}
\alias{structure_ratio_prune}
\title{Tree structure ratio for prune.}
\usage{
structure_ratio_prune(old_tree, tree, current_node, var_in_prune, p_split)
}
\arguments{
\item{old_tree}{The previous tree}

\item{tree}{The current tree}

\item{current_node}{The current pruned node.}

\item{var_in_prune}{The variable that was split in the node chosen to
be pruned.}

\item{p_split}{The number of available predictors}
}
\value{
The tree structure ratio
}
\description{
Tree structure ratio of the candidate tree and
the previous tree, given that the action step was a prune.
}
\details{
For the tree structure ratio of the new pruned tree, we need
to calculate an inversion of the tree structure ratio for the grow.
We need the probabilities of:
1. Splitting at node n
2. Splitting at the node of the left
3. Splitting at the node of the right
4. Using each rule at node n
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
