% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow_ratio.R
\name{transition_ratio_grow}
\alias{transition_ratio_grow}
\title{Transition ratio for grow.}
\usage{
transition_ratio_grow(tree, current_node, p_vars, current_selec_var, p_grow, i)
}
\arguments{
\item{tree}{The current tree}

\item{current_node}{The current node}

\item{p_vars}{The number of predictors still available}

\item{current_selec_var}{The variable selected for the split}

\item{p_grow}{The probability of growing the tree}

\item{i}{The current iteration}
}
\value{
The transition ratio
}
\description{
Transition probability of going to the candidate tree,
given that the action step was a grow
}
\details{
When transitioning to a grow, we need the probabilities of:
 1. Growing the tree
 2. Growing from the specific node, that has to consider the
 available predictors and available values for that grow.
When transitioning from a grow back to the split,
we need the probabilities of:
1. Pruning the tree
2. Selecting  node to prune
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
