% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow_ratio.R
\name{tree_prior}
\alias{tree_prior}
\title{Tree prior}
\usage{
tree_prior(tree, alpha_grow, beta_grow)
}
\arguments{
\item{tree}{The current tree}

\item{alpha_grow}{The alpha of the growing probability}

\item{beta_grow}{The beta of the growing probability}
}
\value{
The tree prior
}
\description{
Calculates the prior for each tree structure
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
