% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow_ratio.R
\name{ratio_grow}
\alias{ratio_grow}
\title{Final ratio for a growth step.}
\usage{
ratio_grow(tree, current_node, pars, p_vars, current_selec_var, i, p_grow)
}
\arguments{
\item{tree}{The current tree}

\item{current_node}{The current grown node}

\item{pars}{The full list of parameters}

\item{p_vars}{The number of available predictors}

\item{current_selec_var}{The variable selected for the split}

\item{i}{The current iteration}

\item{p_grow}{The growing probability}
}
\value{
The final ratio for the candidate tree
}
\description{
The final ratio is to be used as the acceptance
criteria in the MCMC of the hebart model
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
