% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_ratio.R
\name{ratio_prune}
\alias{ratio_prune}
\title{Final ratio for a prune step}
\usage{
ratio_prune(
  tree,
  old_tree,
  current_node,
  pars,
  nodes_to_prune,
  alpha_grow,
  beta_grow
)
}
\arguments{
\item{tree}{The current tree}

\item{old_tree}{The previous tree}

\item{current_node}{The current pruned node}

\item{pars}{The full list of parameters}

\item{nodes_to_prune}{The node to prune from}

\item{alpha_grow}{The alpha of the growing probability}

\item{beta_grow}{The beta of the growing probability}
}
\value{
The final ratio for the candidate tree
}
\description{
The final ratio is to be used as the acceptance
criteria in the MCMC of the b-cart model
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
