% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow_ratio.R
\name{structure_ratio_grow}
\alias{structure_ratio_grow}
\title{Tree structure ratio for grow}
\usage{
structure_ratio_grow(tree, current_node, current_selec_var, p_vars)
}
\arguments{
\item{tree}{The current tree}

\item{current_node}{The current grown node}

\item{current_selec_var}{The variable selected for the split}

\item{p_vars}{The number of available predictors.}
}
\value{
The tree structure ratio
}
\description{
Tree structure ratio of the candidate tree and
the previous tree, given that the action step was a grow
}
\details{
For the tree structure ratio of the new tree, we need
the probabilities of:
}
\author{
Bruna Wundervald, \email{brunadaviesw@gmail.com}.
}
